PositionerSetup: Fix crash fine movement
File "/usr/lib/enigma2/python/Plugins/SystemPlugins/PositionerSetup/ui.py", line 571, in dots
    return (s / 10) * '.' if s < 100 else 10 * '.'
TypeError: can't multiply sequence by non-int of type 'float'

+++ lib/python/Plugins/SystemPlugins/PositionerSetup/ui.py
@@ -568,7 +568,7 @@ class PositionerSetup(Screen):
 	def stepCourse(self, steps):
 		def dots(s):
 			s = abs(s)
-			return (s / 10) * '.' if s < 100 else 10 * '.'
+			return int(s / 10) * '.' if s < 100 else 10 * '.'
 
 		dx = 4 * " "
 		if steps > 0:
